refactor(fault): move to pkg
